Refactor filtering logic to overcome circular referencing. Handle exact string filtering in `filter_result_iterator`.
