+++ include/field.h
@@ -2,13 +2,13 @@
 
 #include <string>
 #include <s2/s2latlng.h>
-#include "art.h"
 #include "option.h"
 #include "string_utils.h"
 #include "logger.h"
 #include "store.h"
 #include <sparsepp.h>
 #include <tsl/htrie_map.h>
+#include <filter.h>
 #include "json.hpp"
 #include "text_embedder_manager.h"
 
@@ -505,200 +505,19 @@ struct field {
     static void compact_nested_fields(tsl::htrie_map<char, field>& nested_fields);
 };
 
-struct filter_node_t;
-
-struct filter {
-    std::string field_name;
-    std::vector<std::string> values;
-    std::vector<NUM_COMPARATOR> comparators;
-    // Would be set when `field: != ...` is encountered with a string field or `field: != [ ... ]` is encountered in the
-    // case of int and float fields. During filtering, all the results of matching the field against the values are
-    // aggregated and then this flag is checked if negation on the aggregated result is required.
-    bool apply_not_equals = false;
-
-    // Would store `Foo` in case of a filter expression like `$Foo(bar := baz)`
-    std::string referenced_collection_name = "";
-
-    static const std::string RANGE_OPERATOR() {
-        return "..";
-    }
-
-    static Option<bool> validate_numerical_filter_value(field _field, const std::string& raw_value) {
-        if(_field.is_int32() && !StringUtils::is_int32_t(raw_value)) {
-            return Option<bool>(400, "Error with filter field `" + _field.name + "`: Not an int32.");
-        }
-
-        else if(_field.is_int64() && !StringUtils::is_int64_t(raw_value)) {
-            return Option<bool>(400, "Error with filter field `" + _field.name + "`: Not an int64.");
-        }
-
-        else if(_field.is_float() && !StringUtils::is_float(raw_value)) {
-            return Option<bool>(400, "Error with filter field `" + _field.name + "`: Not a float.");
-        }
-
-        return Option<bool>(true);
-    }
-
-    static Option<NUM_COMPARATOR> extract_num_comparator(std::string & comp_and_value) {
-        auto num_comparator = EQUALS;
-
-        if(StringUtils::is_integer(comp_and_value) || StringUtils::is_float(comp_and_value)) {
-            num_comparator = EQUALS;
-        }
-
-            // the ordering is important - we have to compare 2-letter operators first
-        else if(comp_and_value.compare(0, 2, "<=") == 0) {
-            num_comparator = LESS_THAN_EQUALS;
-        }
-
-        else if(comp_and_value.compare(0, 2, ">=") == 0) {
-            num_comparator = GREATER_THAN_EQUALS;
-        }
-
-        else if(comp_and_value.compare(0, 2, "!=") == 0) {
-            num_comparator = NOT_EQUALS;
-        }
-
-        else if(comp_and_value.compare(0, 1, "<") == 0) {
-            num_comparator = LESS_THAN;
-        }
-
-        else if(comp_and_value.compare(0, 1, ">") == 0) {
-            num_comparator = GREATER_THAN;
-        }
-
-        else if(comp_and_value.find("..") != std::string::npos) {
-            num_comparator = RANGE_INCLUSIVE;
-        }
-
-        else {
-            return Option<NUM_COMPARATOR>(400, "Numerical field has an invalid comparator.");
-        }
-
-        if(num_comparator == LESS_THAN || num_comparator == GREATER_THAN) {
-            comp_and_value = comp_and_value.substr(1);
-        } else if(num_comparator == LESS_THAN_EQUALS || num_comparator == GREATER_THAN_EQUALS || num_comparator == NOT_EQUALS) {
-            comp_and_value = comp_and_value.substr(2);
-        }
-
-        comp_and_value = StringUtils::trim(comp_and_value);
-
-        return Option<NUM_COMPARATOR>(num_comparator);
-    }
-
-    static Option<bool> parse_geopoint_filter_value(std::string& raw_value,
-                                                    const std::string& format_err_msg,
-                                                    std::string& processed_filter_val,
-                                                    NUM_COMPARATOR& num_comparator);
-
-    static Option<bool> parse_filter_query(const std::string& filter_query,
-                                           const tsl::htrie_map<char, field>& search_schema,
-                                           const Store* store,
-                                           const std::string& doc_id_prefix,
-                                           filter_node_t*& root);
+enum index_operation_t {
+    CREATE,
+    UPSERT,
+    UPDATE,
+    EMPLACE,
+    DELETE
 };
 
-struct filter_node_t {
-    filter filter_exp;
-    FILTER_OPERATOR filter_operator;
-    bool isOperator;
-    filter_node_t* left = nullptr;
-    filter_node_t* right = nullptr;
-
-    filter_node_t(filter filter_exp)
-            : filter_exp(std::move(filter_exp)),
-              isOperator(false),
-              left(nullptr),
-              right(nullptr) {}
-
-    filter_node_t(FILTER_OPERATOR filter_operator,
-                  filter_node_t* left,
-                  filter_node_t* right)
-            : filter_operator(filter_operator),
-              isOperator(true),
-              left(left),
-              right(right) {}
-
-    ~filter_node_t() {
-        delete left;
-        delete right;
-    }
-};
-
-struct reference_filter_result_t {
-    uint32_t count = 0;
-    uint32_t* docs = nullptr;
-
-    reference_filter_result_t& operator=(const reference_filter_result_t& obj) noexcept {
-        if (&obj == this)
-            return *this;
-
-        count = obj.count;
-        docs = new uint32_t[count];
-        memcpy(docs, obj.docs, count * sizeof(uint32_t));
-
-        return *this;
-    }
-
-    ~reference_filter_result_t() {
-        delete[] docs;
-    }
-};
-
-struct filter_result_t {
-    uint32_t count = 0;
-    uint32_t* docs = nullptr;
-    // Collection name -> Reference filter result
-    std::map<std::string, reference_filter_result_t*> reference_filter_results;
-
-    filter_result_t() = default;
-
-    filter_result_t(uint32_t count, uint32_t* docs) : count(count), docs(docs) {}
-
-    filter_result_t& operator=(const filter_result_t& obj) noexcept {
-        if (&obj == this)
-            return *this;
-
-        count = obj.count;
-        docs = new uint32_t[count];
-        memcpy(docs, obj.docs, count * sizeof(uint32_t));
-
-        // Copy every collection's references.
-        for (const auto &item: obj.reference_filter_results) {
-            reference_filter_results[item.first] = new reference_filter_result_t[count];
-
-            for (uint32_t i = 0; i < count; i++) {
-                reference_filter_results[item.first][i] = item.second[i];
-            }
-        }
-
-        return *this;
-    }
-
-    filter_result_t& operator=(filter_result_t&& obj) noexcept {
-        if (&obj == this)
-            return *this;
-
-        count = obj.count;
-        docs = obj.docs;
-        reference_filter_results = std::map(obj.reference_filter_results);
-
-        obj.docs = nullptr;
-        obj.reference_filter_results.clear();
-
-        return *this;
-    }
-
-    ~filter_result_t() {
-        delete[] docs;
-        for (const auto &item: reference_filter_results) {
-            delete[] item.second;
-        }
-    }
-
-    static void and_filter_results(const filter_result_t& a, const filter_result_t& b, filter_result_t& result);
-
-    static void or_filter_results(const filter_result_t& a, const filter_result_t& b, filter_result_t& result);
+enum class DIRTY_VALUES {
+    REJECT = 1,
+    DROP = 2,
+    COERCE_OR_REJECT = 3,
+    COERCE_OR_DROP = 4,
 };
 
 namespace sort_field_const {
