+++ src/art.cpp
@@ -22,6 +22,7 @@
 #include "art.h"
 #include "logger.h"
 #include "array_utils.h"
+#include "filter_result_iterator.h"
 
 /**
  * Macros to manipulate pointer tags
@@ -972,6 +973,36 @@ const uint32_t* get_allowed_doc_ids(art_tree *t, const std::string& prev_token,
     return prev_token_doc_ids;
 }
 
+const uint32_t* get_allowed_doc_ids(art_tree *t, const std::string& prev_token,
+                                    filter_result_iterator_t& filter_result_iterator,
+                                    size_t& prev_token_doc_ids_len) {
+
+    art_leaf* prev_leaf = static_cast<art_leaf*>(
+            art_search(t, reinterpret_cast<const unsigned char*>(prev_token.c_str()), prev_token.size() + 1)
+    );
+
+    uint32_t* prev_token_doc_ids = nullptr;
+
+    if(prev_token.empty() || !prev_leaf) {
+        prev_token_doc_ids_len = filter_result_iterator.to_filter_id_array(prev_token_doc_ids);
+        return prev_token_doc_ids;
+    }
+
+    std::vector<uint32_t> prev_leaf_ids;
+    posting_t::merge({prev_leaf->values}, prev_leaf_ids);
+
+    if(filter_result_iterator.valid()) {
+        prev_token_doc_ids_len = filter_result_iterator.and_scalar(prev_leaf_ids.data(), prev_leaf_ids.size(),
+                                                                   prev_token_doc_ids);
+    } else {
+        prev_token_doc_ids_len = prev_leaf_ids.size();
+        prev_token_doc_ids = new uint32_t[prev_token_doc_ids_len];
+        std::copy(prev_leaf_ids.begin(), prev_leaf_ids.end(), prev_token_doc_ids);
+    }
+
+    return prev_token_doc_ids;
+}
+
 bool validate_and_add_leaf(art_leaf* leaf, const bool last_token, const std::string& prev_token,
                            const uint32_t* allowed_doc_ids, const size_t allowed_doc_ids_len,
                            std::set<std::string>& exclude_leaves, const art_leaf* exact_leaf,
@@ -1618,6 +1649,165 @@ int art_fuzzy_search(art_tree *t, const unsigned char *term, const int term_len,
     if(allowed_doc_ids != filter_ids) {
         delete [] allowed_doc_ids;
     }
+
+    return 0;
+}
+
+int art_fuzzy_search_i(art_tree *t, const unsigned char *term, const int term_len, const int min_cost, const int max_cost,
+                       const size_t max_words, const token_ordering token_order,
+                       const bool prefix, bool last_token, const std::string& prev_token,
+                       filter_result_iterator_t& filter_result_iterator,
+                       std::vector<art_leaf *> &results, std::set<std::string>& exclude_leaves) {
+
+    std::vector<const art_node*> nodes;
+    int irow[term_len + 1];
+    int jrow[term_len + 1];
+    for (int i = 0; i <= term_len; i++){
+        irow[i] = jrow[i] = i;
+    }
+
+    //auto begin = std::chrono::high_resolution_clock::now();
+
+    if(IS_LEAF(t->root)) {
+        art_leaf *l = (art_leaf *) LEAF_RAW(t->root);
+        art_fuzzy_recurse(0, l->key[0], t->root, 0, term, term_len, irow, jrow, min_cost, max_cost, prefix, nodes);
+    } else {
+        if(t->root == nullptr) {
+            return 0;
+        }
+
+        // send depth as -1 to indicate that this is a root node
+        art_fuzzy_recurse(0, 0, t->root, -1, term, term_len, irow, jrow, min_cost, max_cost, prefix, nodes);
+    }
+
+    //long long int time_micro = microseconds(std::chrono::high_resolution_clock::now() - begin).count();
+    //!LOG(INFO) << "Time taken for fuzz: " << time_micro << "us, size of nodes: " << nodes.size();
+
+    //auto begin = std::chrono::high_resolution_clock::now();
+
+    size_t key_len = prefix ? term_len + 1 : term_len;
+    art_leaf* exact_leaf = (art_leaf *) art_search(t, term, key_len);
+    //LOG(INFO) << "exact_leaf: " << exact_leaf << ", term: " << term << ", term_len: " << term_len;
+
+    // documents that contain the previous token and/or filter ids
+    size_t allowed_doc_ids_len = 0;
+    const uint32_t* allowed_doc_ids = get_allowed_doc_ids(t, prev_token, filter_result_iterator,  allowed_doc_ids_len);
+
+    for(auto node: nodes) {
+        art_topk_iter(node, token_order, max_words, exact_leaf,
+                      last_token, prev_token, allowed_doc_ids, allowed_doc_ids_len,
+                      t, exclude_leaves, results);
+    }
+
+    if(token_order == FREQUENCY) {
+        std::sort(results.begin(), results.end(), compare_art_leaf_frequency);
+    } else {
+        std::sort(results.begin(), results.end(), compare_art_leaf_score);
+    }
+
+    if(exact_leaf && min_cost == 0) {
+        std::string tok(reinterpret_cast<char*>(exact_leaf->key), exact_leaf->key_len - 1);
+        if(exclude_leaves.count(tok) == 0) {
+            results.insert(results.begin(), exact_leaf);
+            exclude_leaves.emplace(tok);
+        }
+    }
+
+    if(results.size() > max_words) {
+        results.resize(max_words);
+    }
+
+    /*auto time_micro = microseconds(std::chrono::high_resolution_clock::now() - begin).count();
+    if(time_micro > 1000) {
+        LOG(INFO) << "Time taken for art_topk_iter: " << time_micro
+                  << "us, size of nodes: " << nodes.size()
+                  << ", filter_ids_length: " << filter_ids_length;
+    }*/
+
+// TODO: Figure out this edge case.
+//    if(allowed_doc_ids != filter_ids) {
+//        delete [] allowed_doc_ids;
+//    }
+
+    return 0;
+}
+
+int art_fuzzy_search(art_tree *t, const unsigned char *term, const int term_len, const int min_cost, const int max_cost,
+                     const size_t max_words, const token_ordering token_order, const bool prefix,
+                     bool last_token, const std::string& prev_token,
+                     filter_result_iterator_t& filter_result_iterator,
+                     std::vector<art_leaf *> &results, std::set<std::string>& exclude_leaves) {
+
+    std::vector<const art_node*> nodes;
+    int irow[term_len + 1];
+    int jrow[term_len + 1];
+    for (int i = 0; i <= term_len; i++){
+        irow[i] = jrow[i] = i;
+    }
+
+    //auto begin = std::chrono::high_resolution_clock::now();
+
+    if(IS_LEAF(t->root)) {
+        art_leaf *l = (art_leaf *) LEAF_RAW(t->root);
+        art_fuzzy_recurse(0, l->key[0], t->root, 0, term, term_len, irow, jrow, min_cost, max_cost, prefix, nodes);
+    } else {
+        if(t->root == nullptr) {
+            return 0;
+        }
+
+        // send depth as -1 to indicate that this is a root node
+        art_fuzzy_recurse(0, 0, t->root, -1, term, term_len, irow, jrow, min_cost, max_cost, prefix, nodes);
+    }
+
+    //long long int time_micro = microseconds(std::chrono::high_resolution_clock::now() - begin).count();
+    //!LOG(INFO) << "Time taken for fuzz: " << time_micro << "us, size of nodes: " << nodes.size();
+
+    //auto begin = std::chrono::high_resolution_clock::now();
+
+    size_t key_len = prefix ? term_len + 1 : term_len;
+    art_leaf* exact_leaf = (art_leaf *) art_search(t, term, key_len);
+    //LOG(INFO) << "exact_leaf: " << exact_leaf << ", term: " << term << ", term_len: " << term_len;
+
+    // documents that contain the previous token and/or filter ids
+    size_t allowed_doc_ids_len = 0;
+    const uint32_t* allowed_doc_ids = get_allowed_doc_ids(t, prev_token, filter_result_iterator,  allowed_doc_ids_len);
+
+    for(auto node: nodes) {
+        art_topk_iter(node, token_order, max_words, exact_leaf,
+                      last_token, prev_token, allowed_doc_ids, allowed_doc_ids_len,
+                      t, exclude_leaves, results);
+    }
+
+    if(token_order == FREQUENCY) {
+        std::sort(results.begin(), results.end(), compare_art_leaf_frequency);
+    } else {
+        std::sort(results.begin(), results.end(), compare_art_leaf_score);
+    }
+
+    if(exact_leaf && min_cost == 0) {
+        std::string tok(reinterpret_cast<char*>(exact_leaf->key), exact_leaf->key_len - 1);
+        if(exclude_leaves.count(tok) == 0) {
+            results.insert(results.begin(), exact_leaf);
+            exclude_leaves.emplace(tok);
+        }
+    }
+
+    if(results.size() > max_words) {
+        results.resize(max_words);
+    }
+
+    /*auto time_micro = microseconds(std::chrono::high_resolution_clock::now() - begin).count();
+
+    if(time_micro > 1000) {
+        LOG(INFO) << "Time taken for art_topk_iter: " << time_micro
+                  << "us, size of nodes: " << nodes.size()
+                  << ", filter_ids_length: " << filter_ids_length;
+    }*/
+
+// TODO: Figure out this edge case.
+//    if(allowed_doc_ids != filter_ids) {
+//        delete [] allowed_doc_ids;
+//    }
 
     return 0;
 }
