+++ src/field.cpp
@@ -6,511 +6,6 @@
 #include <collection_manager.h>
 #include <regex>
 
-Option<bool> filter::parse_geopoint_filter_value(std::string& raw_value,
-                                                 const std::string& format_err_msg,
-                                                 std::string& processed_filter_val,
-                                                 NUM_COMPARATOR& num_comparator) {
-
-    num_comparator = LESS_THAN_EQUALS;
-
-    if(!(raw_value[0] == '(' && raw_value[raw_value.size() - 1] == ')')) {
-        return Option<bool>(400, format_err_msg);
-    }
-
-    std::vector<std::string> filter_values;
-    auto raw_val_without_paran = raw_value.substr(1, raw_value.size() - 2);
-    StringUtils::split(raw_val_without_paran, filter_values, ",");
-
-    // we will end up with: "10.45 34.56 2 km" or "10.45 34.56 2mi" or a geo polygon
-
-    if(filter_values.size() < 3) {
-        return Option<bool>(400, format_err_msg);
-    }
-
-    // do validation: format should match either a point + radius or polygon
-
-    size_t num_floats = 0;
-    for(const auto& fvalue: filter_values) {
-        if(StringUtils::is_float(fvalue)) {
-            num_floats++;
-        }
-    }
-
-    bool is_polygon = (num_floats == filter_values.size());
-    if(!is_polygon) {
-        // we have to ensure that this is a point + radius match
-        if(!StringUtils::is_float(filter_values[0]) || !StringUtils::is_float(filter_values[1])) {
-            return Option<bool>(400, format_err_msg);
-        }
-
-        if(filter_values[0] == "nan" || filter_values[0] == "NaN" ||
-            filter_values[1] == "nan" || filter_values[1] == "NaN") {
-            return Option<bool>(400, format_err_msg);
-        }
-    }
-
-    if(is_polygon) {
-        processed_filter_val = raw_val_without_paran;
-    } else {
-        // point + radius
-        // filter_values[2] is distance, get the unit, validate it and split on that
-        if(filter_values[2].size() < 2) {
-            return Option<bool>(400, "Unit must be either `km` or `mi`.");
-        }
-
-        std::string unit = filter_values[2].substr(filter_values[2].size()-2, 2);
-
-        if(unit != "km" && unit != "mi") {
-            return Option<bool>(400, "Unit must be either `km` or `mi`.");
-        }
-
-        std::vector<std::string> dist_values;
-        StringUtils::split(filter_values[2], dist_values, unit);
-
-        if(dist_values.size() != 1) {
-            return Option<bool>(400, format_err_msg);
-        }
-
-        if(!StringUtils::is_float(dist_values[0])) {
-            return Option<bool>(400, format_err_msg);
-        }
-
-        processed_filter_val = filter_values[0] + ", " + filter_values[1] + ", " + // co-ords
-                               dist_values[0] + ", " +  unit;           // X km
-    }
-
-    return Option<bool>(true);
-}
-
-bool isOperator(const std::string& expression) {
-    return expression == "&&" || expression == "||";
-}
-
-// https://en.wikipedia.org/wiki/Shunting_yard_algorithm
-Option<bool> toPostfix(std::queue<std::string>& tokens, std::queue<std::string>& postfix) {
-    std::stack<std::string> operatorStack;
-
-    while (!tokens.empty()) {
-        auto expression = tokens.front();
-        tokens.pop();
-
-        if (isOperator(expression)) {
-            // We only have two operators &&, || having the same precedence and both being left associative.
-            while (!operatorStack.empty() && operatorStack.top() != "(") {
-                postfix.push(operatorStack.top());
-                operatorStack.pop();
-            }
-
-            operatorStack.push(expression);
-        } else if (expression == "(") {
-            operatorStack.push(expression);
-        } else if (expression == ")") {
-            while (!operatorStack.empty() && operatorStack.top() != "(") {
-                postfix.push(operatorStack.top());
-                operatorStack.pop();
-            }
-
-            if (operatorStack.empty() || operatorStack.top() != "(") {
-                return Option<bool>(400, "Could not parse the filter query: unbalanced parentheses.");
-            }
-            operatorStack.pop();
-        } else {
-            postfix.push(expression);
-        }
-    }
-
-    while (!operatorStack.empty()) {
-        if (operatorStack.top() == "(") {
-            return Option<bool>(400, "Could not parse the filter query: unbalanced parentheses.");
-        }
-        postfix.push(operatorStack.top());
-        operatorStack.pop();
-    }
-
-    return Option<bool>(true);
-}
-
-Option<bool> toMultiValueNumericFilter(std::string& raw_value, filter& filter_exp, const field& _field) {
-    std::vector<std::string> filter_values;
-    StringUtils::split(raw_value.substr(1, raw_value.size() - 2), filter_values, ",");
-    filter_exp = {_field.name, {}, {}};
-    for (std::string& filter_value: filter_values) {
-        Option<NUM_COMPARATOR> op_comparator = filter::extract_num_comparator(filter_value);
-        if (!op_comparator.ok()) {
-            return Option<bool>(400, "Error with filter field `" + _field.name + "`: " + op_comparator.error());
-        }
-        if (op_comparator.get() == RANGE_INCLUSIVE) {
-            // split the value around range operator to extract bounds
-            std::vector<std::string> range_values;
-            StringUtils::split(filter_value, range_values, filter::RANGE_OPERATOR());
-            for (const std::string& range_value: range_values) {
-                auto validate_op = filter::validate_numerical_filter_value(_field, range_value);
-                if (!validate_op.ok()) {
-                    return validate_op;
-                }
-                filter_exp.values.push_back(range_value);
-                filter_exp.comparators.push_back(op_comparator.get());
-            }
-        } else {
-            auto validate_op = filter::validate_numerical_filter_value(_field, filter_value);
-            if (!validate_op.ok()) {
-                return validate_op;
-            }
-            filter_exp.values.push_back(filter_value);
-            filter_exp.comparators.push_back(op_comparator.get());
-        }
-    }
-
-    return Option<bool>(true);
-}
-
-Option<bool> toFilter(const std::string expression,
-                      filter& filter_exp,
-                      const tsl::htrie_map<char, field>& search_schema,
-                      const Store* store,
-                      const std::string& doc_id_prefix) {
-    // split into [field_name, value]
-    size_t found_index = expression.find(':');
-    if (found_index == std::string::npos) {
-        return Option<bool>(400, "Could not parse the filter query.");
-    }
-    std::string&& field_name = expression.substr(0, found_index);
-    StringUtils::trim(field_name);
-    if (field_name == "id") {
-        std::string&& raw_value = expression.substr(found_index + 1, std::string::npos);
-        StringUtils::trim(raw_value);
-        std::string empty_filter_err = "Error with filter field `id`: Filter value cannot be empty.";
-        if (raw_value.empty()) {
-            return Option<bool>(400, empty_filter_err);
-        }
-        filter_exp = {field_name, {}, {}};
-        NUM_COMPARATOR id_comparator = EQUALS;
-        size_t filter_value_index = 0;
-        if (raw_value[0] == '=') {
-            id_comparator = EQUALS;
-            while (++filter_value_index < raw_value.size() && raw_value[filter_value_index] == ' ');
-        } else if (raw_value.size() >= 2 && raw_value[0] == '!' && raw_value[1] == '=') {
-            return Option<bool>(400, "Not equals filtering is not supported on the `id` field.");
-        }
-        if (filter_value_index != 0) {
-            raw_value = raw_value.substr(filter_value_index);
-        }
-        if (raw_value.empty()) {
-            return Option<bool>(400, empty_filter_err);
-        }
-        if (raw_value[0] == '[' && raw_value[raw_value.size() - 1] == ']') {
-            std::vector<std::string> doc_ids;
-            StringUtils::split_to_values(raw_value.substr(1, raw_value.size() - 2), doc_ids);
-            for (std::string& doc_id: doc_ids) {
-                // we have to convert the doc_id to seq id
-                std::string seq_id_str;
-                StoreStatus seq_id_status = store->get(doc_id_prefix + doc_id, seq_id_str);
-                if (seq_id_status != StoreStatus::FOUND) {
-                    continue;
-                }
-                filter_exp.values.push_back(seq_id_str);
-                filter_exp.comparators.push_back(id_comparator);
-            }
-        } else {
-            std::vector<std::string> doc_ids;
-            StringUtils::split_to_values(raw_value, doc_ids); // to handle backticks
-            std::string seq_id_str;
-            StoreStatus seq_id_status = store->get(doc_id_prefix + doc_ids[0], seq_id_str);
-            if (seq_id_status == StoreStatus::FOUND) {
-                filter_exp.values.push_back(seq_id_str);
-                filter_exp.comparators.push_back(id_comparator);
-            }
-        }
-        return Option<bool>(true);
-    }
-
-    auto field_it = search_schema.find(field_name);
-
-    if (field_it == search_schema.end()) {
-        return Option<bool>(404, "Could not find a filter field named `" + field_name + "` in the schema.");
-    }
-
-    if (field_it->num_dim > 0) {
-        return Option<bool>(404, "Cannot filter on vector field `" + field_name + "`.");
-    }
-
-    const field& _field = field_it.value();
-    std::string&& raw_value = expression.substr(found_index + 1, std::string::npos);
-    StringUtils::trim(raw_value);
-    // skip past optional `:=` operator, which has no meaning for non-string fields
-    if (!_field.is_string() && raw_value[0] == '=') {
-        size_t filter_value_index = 0;
-        while (raw_value[++filter_value_index] == ' ');
-        raw_value = raw_value.substr(filter_value_index);
-    }
-    if (_field.is_integer() || _field.is_float()) {
-        // could be a single value or a list
-        if (raw_value[0] == '[' && raw_value[raw_value.size() - 1] == ']') {
-            Option<bool> op = toMultiValueNumericFilter(raw_value, filter_exp, _field);
-            if (!op.ok()) {
-                return op;
-            }
-        } else {
-            Option<NUM_COMPARATOR> op_comparator = filter::extract_num_comparator(raw_value);
-            if (!op_comparator.ok()) {
-                return Option<bool>(400, "Error with filter field `" + _field.name + "`: " + op_comparator.error());
-            }
-            if (op_comparator.get() == RANGE_INCLUSIVE) {
-                // split the value around range operator to extract bounds
-                std::vector<std::string> range_values;
-                StringUtils::split(raw_value, range_values, filter::RANGE_OPERATOR());
-                filter_exp.field_name = field_name;
-                for (const std::string& range_value: range_values) {
-                    auto validate_op = filter::validate_numerical_filter_value(_field, range_value);
-                    if (!validate_op.ok()) {
-                        return validate_op;
-                    }
-                    filter_exp.values.push_back(range_value);
-                    filter_exp.comparators.push_back(op_comparator.get());
-                }
-            } else if (op_comparator.get() == NOT_EQUALS && raw_value[0] == '[' && raw_value[raw_value.size() - 1] == ']') {
-                Option<bool> op = toMultiValueNumericFilter(raw_value, filter_exp, _field);
-                if (!op.ok()) {
-                    return op;
-                }
-                filter_exp.apply_not_equals = true;
-            } else {
-                auto validate_op = filter::validate_numerical_filter_value(_field, raw_value);
-                if (!validate_op.ok()) {
-                    return validate_op;
-                }
-                filter_exp = {field_name, {raw_value}, {op_comparator.get()}};
-            }
-        }
-    } else if (_field.is_bool()) {
-        NUM_COMPARATOR bool_comparator = EQUALS;
-        size_t filter_value_index = 0;
-        if (raw_value[0] == '=') {
-            bool_comparator = EQUALS;
-            while (++filter_value_index < raw_value.size() && raw_value[filter_value_index] == ' ');
-        } else if (raw_value.size() >= 2 && raw_value[0] == '!' && raw_value[1] == '=') {
-            bool_comparator = NOT_EQUALS;
-            filter_value_index++;
-            while (++filter_value_index < raw_value.size() && raw_value[filter_value_index] == ' ');
-        }
-        if (filter_value_index != 0) {
-            raw_value = raw_value.substr(filter_value_index);
-        }
-        if (filter_value_index == raw_value.size()) {
-            return Option<bool>(400, "Error with filter field `" + _field.name +
-                                     "`: Filter value cannot be empty.");
-        }
-        if (raw_value[0] == '[' && raw_value[raw_value.size() - 1] == ']') {
-            std::vector<std::string> filter_values;
-            StringUtils::split(raw_value.substr(1, raw_value.size() - 2), filter_values, ",");
-            filter_exp = {field_name, {}, {}};
-            for (std::string& filter_value: filter_values) {
-                if (filter_value != "true" && filter_value != "false") {
-                    return Option<bool>(400, "Values of filter field `" + _field.name +
-                                             "`: must be `true` or `false`.");
-                }
-                filter_value = (filter_value == "true") ? "1" : "0";
-                filter_exp.values.push_back(filter_value);
-                filter_exp.comparators.push_back(bool_comparator);
-            }
-        } else {
-            if (raw_value != "true" && raw_value != "false") {
-                return Option<bool>(400, "Value of filter field `" + _field.name + "` must be `true` or `false`.");
-            }
-            std::string bool_value = (raw_value == "true") ? "1" : "0";
-            filter_exp = {field_name, {bool_value}, {bool_comparator}};
-        }
-    } else if (_field.is_geopoint()) {
-        filter_exp = {field_name, {}, {}};
-        const std::string& format_err_msg = "Value of filter field `" + _field.name +
-                                            "`: must be in the `(-44.50, 170.29, 0.75 km)` or "
-                                            "(56.33, -65.97, 23.82, -127.82) format.";
-        NUM_COMPARATOR num_comparator;
-        // could be a single value or a list
-        if (raw_value[0] == '[' && raw_value[raw_value.size() - 1] == ']') {
-            std::vector<std::string> filter_values;
-            StringUtils::split(raw_value.substr(1, raw_value.size() - 2), filter_values, "),");
-            for (std::string& filter_value: filter_values) {
-                filter_value += ")";
-                std::string processed_filter_val;
-                auto parse_op = filter::parse_geopoint_filter_value(filter_value, format_err_msg, processed_filter_val,
-                                                                    num_comparator);
-                if (!parse_op.ok()) {
-                    return parse_op;
-                }
-                filter_exp.values.push_back(processed_filter_val);
-                filter_exp.comparators.push_back(num_comparator);
-            }
-        } else {
-            // single value, e.g. (10.45, 34.56, 2 km)
-            std::string processed_filter_val;
-            auto parse_op = filter::parse_geopoint_filter_value(raw_value, format_err_msg, processed_filter_val,
-                                                                num_comparator);
-            if (!parse_op.ok()) {
-                return parse_op;
-            }
-            filter_exp.values.push_back(processed_filter_val);
-            filter_exp.comparators.push_back(num_comparator);
-        }
-    } else if (_field.is_string()) {
-        size_t filter_value_index = 0;
-        NUM_COMPARATOR str_comparator = CONTAINS;
-        if (raw_value[0] == '=') {
-            // string filter should be evaluated in strict "equals" mode
-            str_comparator = EQUALS;
-            while (++filter_value_index < raw_value.size() && raw_value[filter_value_index] == ' ');
-        } else if (raw_value.size() >= 2 && raw_value[0] == '!' && raw_value[1] == '=') {
-            str_comparator = NOT_EQUALS;
-            filter_value_index++;
-            while (++filter_value_index < raw_value.size() && raw_value[filter_value_index] == ' ');
-        }
-        if (filter_value_index == raw_value.size()) {
-            return Option<bool>(400, "Error with filter field `" + _field.name +
-                                     "`: Filter value cannot be empty.");
-        }
-        if (raw_value[filter_value_index] == '[' && raw_value[raw_value.size() - 1] == ']') {
-            std::vector<std::string> filter_values;
-            StringUtils::split_to_values(
-                    raw_value.substr(filter_value_index + 1, raw_value.size() - filter_value_index - 2), filter_values);
-            filter_exp = {field_name, filter_values, {str_comparator}};
-        } else {
-            filter_exp = {field_name, {raw_value.substr(filter_value_index)}, {str_comparator}};
-        }
-
-        filter_exp.apply_not_equals = (str_comparator == NOT_EQUALS);
-    } else {
-        return Option<bool>(400, "Error with filter field `" + _field.name +
-                                 "`: Unidentified field data type, see docs for supported data types.");
-    }
-
-    return Option<bool>(true);
-}
-
-// https://stackoverflow.com/a/423914/11218270
-Option<bool> toParseTree(std::queue<std::string>& postfix, filter_node_t*& root,
-                         const tsl::htrie_map<char, field>& search_schema,
-                         const Store* store,
-                         const std::string& doc_id_prefix) {
-    std::stack<filter_node_t*> nodeStack;
-    bool is_successful = true;
-    std::string error_message;
-
-    filter_node_t *filter_node = nullptr;
-
-    while (!postfix.empty()) {
-        const std::string expression = postfix.front();
-        postfix.pop();
-
-        if (isOperator(expression)) {
-            if (nodeStack.empty()) {
-                is_successful = false;
-                error_message = "Could not parse the filter query: unbalanced `" + expression + "` operands.";
-                break;
-            }
-            auto operandB = nodeStack.top();
-            nodeStack.pop();
-
-            if (nodeStack.empty()) {
-                delete operandB;
-                is_successful = false;
-                error_message = "Could not parse the filter query: unbalanced `" + expression + "` operands.";
-                break;
-            }
-            auto operandA = nodeStack.top();
-            nodeStack.pop();
-
-            filter_node = new filter_node_t(expression == "&&" ? AND : OR, operandA, operandB);
-        } else {
-            filter filter_exp;
-
-            // Expected value: $Collection(...)
-            bool is_referenced_filter = (expression[0] == '$' && expression[expression.size() - 1] == ')');
-            if (is_referenced_filter) {
-                size_t parenthesis_index = expression.find('(');
-
-                std::string collection_name = expression.substr(1, parenthesis_index - 1);
-                auto &cm = CollectionManager::get_instance();
-                auto collection = cm.get_collection(collection_name);
-                if (collection == nullptr) {
-                    is_successful = false;
-                    error_message = "Referenced collection `" + collection_name + "` not found.";
-                    break;
-                }
-
-                filter_exp = {expression.substr(parenthesis_index + 1, expression.size() - parenthesis_index - 2)};
-                filter_exp.referenced_collection_name = collection_name;
-            } else {
-                Option<bool> toFilter_op = toFilter(expression, filter_exp, search_schema, store, doc_id_prefix);
-                if (!toFilter_op.ok()) {
-                    is_successful = false;
-                    error_message = toFilter_op.error();
-                    break;
-                }
-            }
-
-            filter_node = new filter_node_t(filter_exp);
-        }
-
-        nodeStack.push(filter_node);
-    }
-
-    if (!is_successful) {
-        while (!nodeStack.empty()) {
-            auto filterNode = nodeStack.top();
-            delete filterNode;
-            nodeStack.pop();
-        }
-
-        return Option<bool>(400, error_message);
-    }
-
-    if (nodeStack.empty()) {
-        return Option<bool>(400, "Filter query cannot be empty.");
-    }
-    root = nodeStack.top();
-
-    return Option<bool>(true);
-}
-
-Option<bool> filter::parse_filter_query(const std::string& filter_query,
-                                        const tsl::htrie_map<char, field>& search_schema,
-                                        const Store* store,
-                                        const std::string& doc_id_prefix,
-                                        filter_node_t*& root) {
-    auto _filter_query = filter_query;
-    StringUtils::trim(_filter_query);
-    if (_filter_query.empty()) {
-        return Option<bool>(true);
-    }
-
-    std::queue<std::string> tokens;
-    Option<bool> tokenize_op = StringUtils::tokenize_filter_query(filter_query, tokens);
-    if (!tokenize_op.ok()) {
-        return tokenize_op;
-    }
-
-    if (tokens.size() > 100) {
-        return Option<bool>(400, "Filter expression is not valid.");
-    }
-
-    std::queue<std::string> postfix;
-    Option<bool> toPostfix_op = toPostfix(tokens, postfix);
-    if (!toPostfix_op.ok()) {
-        return toPostfix_op;
-    }
-
-    Option<bool> toParseTree_op = toParseTree(postfix,
-                                              root,
-                                              search_schema,
-                                              store,
-                                              doc_id_prefix);
-    if (!toParseTree_op.ok()) {
-        return toParseTree_op;
-    }
-
-    return Option<bool>(true);
-}
-
 Option<bool> field::json_field_to_field(bool enable_nested_fields, nlohmann::json& field_json,
                                         std::vector<field>& the_fields,
                                         string& fallback_field_type, size_t& num_auto_detect_fields) {
@@ -1028,169 +523,5 @@ void field::compact_nested_fields(tsl::htrie_map<char, field>& nested_fields) {
 
     for(auto& field_name: nested_fields_vec) {
         nested_fields.erase_prefix(field_name + ".");
-    }
-}
-
-void filter_result_t::and_filter_results(const filter_result_t& a, const filter_result_t& b, filter_result_t& result) {
-    auto lenA = a.count, lenB = b.count;
-    if (lenA == 0 || lenB == 0) {
-        return;
-    }
-
-    result.docs = new uint32_t[std::min(lenA, lenB)];
-
-    auto A = a.docs, B = b.docs, out = result.docs;
-    const uint32_t *endA = A + lenA;
-    const uint32_t *endB = B + lenB;
-
-    // Add an entry of references in the result for each unique collection in a and b.
-    for (auto const& item: a.reference_filter_results) {
-        if (result.reference_filter_results.count(item.first) == 0) {
-            result.reference_filter_results[item.first] = new reference_filter_result_t[std::min(lenA, lenB)];
-        }
-    }
-    for (auto const& item: b.reference_filter_results) {
-        if (result.reference_filter_results.count(item.first) == 0) {
-            result.reference_filter_results[item.first] = new reference_filter_result_t[std::min(lenA, lenB)];
-        }
-    }
-
-    while (true) {
-        while (*A < *B) {
-            SKIP_FIRST_COMPARE:
-            if (++A == endA) {
-                result.count = out - result.docs;
-                return;
-            }
-        }
-        while (*A > *B) {
-            if (++B == endB) {
-                result.count = out - result.docs;
-                return;
-            }
-        }
-        if (*A == *B) {
-            *out = *A;
-
-            // Copy the references of the document from every collection into result.
-            for (auto const& item: a.reference_filter_results) {
-                result.reference_filter_results[item.first][out - result.docs] = item.second[A - a.docs];
-            }
-            for (auto const& item: b.reference_filter_results) {
-                result.reference_filter_results[item.first][out - result.docs] = item.second[B - b.docs];
-            }
-
-            out++;
-
-            if (++A == endA || ++B == endB) {
-                result.count = out - result.docs;
-                return;
-            }
-        } else {
-            goto SKIP_FIRST_COMPARE;
-        }
-    }
-}
-
-void filter_result_t::or_filter_results(const filter_result_t& a, const filter_result_t& b, filter_result_t& result) {
-    if (a.count == 0 && b.count == 0) {
-        return;
-    }
-
-    // If either one of a or b does not have any matches, copy other into result.
-    if (a.count == 0) {
-        result = b;
-        return;
-    }
-    if (b.count == 0) {
-        result = a;
-        return;
-    }
-
-    size_t indexA = 0, indexB = 0, res_index = 0, lenA = a.count, lenB = b.count;
-    result.docs = new uint32_t[lenA + lenB];
-
-    // Add an entry of references in the result for each unique collection in a and b.
-    for (auto const& item: a.reference_filter_results) {
-        if (result.reference_filter_results.count(item.first) == 0) {
-            result.reference_filter_results[item.first] = new reference_filter_result_t[lenA + lenB];
-        }
-    }
-    for (auto const& item: b.reference_filter_results) {
-        if (result.reference_filter_results.count(item.first) == 0) {
-            result.reference_filter_results[item.first] = new reference_filter_result_t[lenA + lenB];
-        }
-    }
-
-    while (indexA < lenA && indexB < lenB) {
-        if (a.docs[indexA] < b.docs[indexB]) {
-            // check for duplicate
-            if (res_index == 0 || result.docs[res_index - 1] != a.docs[indexA]) {
-                result.docs[res_index] = a.docs[indexA];
-                res_index++;
-            }
-
-            // Copy references of the last result document from every collection in a.
-            for (auto const& item: a.reference_filter_results) {
-                result.reference_filter_results[item.first][res_index - 1] = item.second[indexA];
-            }
-
-            indexA++;
-        } else {
-            if (res_index == 0 || result.docs[res_index - 1] != b.docs[indexB]) {
-                result.docs[res_index] = b.docs[indexB];
-                res_index++;
-            }
-
-            for (auto const& item: b.reference_filter_results) {
-                result.reference_filter_results[item.first][res_index - 1] = item.second[indexB];
-            }
-
-            indexB++;
-        }
-    }
-
-    while (indexA < lenA) {
-        if (res_index == 0 || result.docs[res_index - 1] != a.docs[indexA]) {
-            result.docs[res_index] = a.docs[indexA];
-            res_index++;
-        }
-
-        for (auto const& item: a.reference_filter_results) {
-            result.reference_filter_results[item.first][res_index - 1] = item.second[indexA];
-        }
-
-        indexA++;
-    }
-
-    while (indexB < lenB) {
-        if(res_index == 0 || result.docs[res_index - 1] != b.docs[indexB]) {
-            result.docs[res_index] = b.docs[indexB];
-            res_index++;
-        }
-
-        for (auto const& item: b.reference_filter_results) {
-            result.reference_filter_results[item.first][res_index - 1] = item.second[indexB];
-        }
-
-        indexB++;
-    }
-
-    result.count = res_index;
-
-    // shrink fit
-    auto out = new uint32_t[res_index];
-    memcpy(out, result.docs, res_index * sizeof(uint32_t));
-    delete[] result.docs;
-    result.docs = out;
-
-    for (auto &item: result.reference_filter_results) {
-        auto out_references = new reference_filter_result_t[res_index];
-
-        for (uint32_t i = 0; i < result.count; i++) {
-            out_references[i] = item.second[i];
-        }
-        delete[] item.second;
-        item.second = out_references;
     }
 }
