+++ src/or_iterator.cpp
@@ -1,5 +1,5 @@
 #include "or_iterator.h"
-
+#include "filter.h"
 
 bool or_iterator_t::at_end(const std::vector<or_iterator_t>& its) {
     // if any iterator is invalid, we stop
@@ -206,6 +206,10 @@ bool or_iterator_t::take_id(result_iter_state_t& istate, uint32_t id, bool& is_e
         }
 
         return false;
+    }
+
+    if (istate.fit != nullptr) {
+        return (istate.fit->valid(id) == 1);
     }
 
     return true;
