+++ test/collection_test.cpp
@@ -5,6 +5,7 @@
 #include <algorithm>
 #include <filesystem>
 #include <collection_manager.h>
+#include <validator.h>
 #include "collection.h"
 #include "text_embedder_manager.h"
 #include "http_client.h"
